I did not need to change the servo library.
This branch uses the same servo library, but I have fixed an array overflow in the scan code.  It was not properly resetting the maxScan index amount.  It was letting the index get to 30 instead of 29, and rewrite some random important thing in the memory.

Anway, this code works and will be migrated over to the main branch.